14A - Claude
Fase 14 (Economy, Jobs & Businesses): fundação da moeda.

Conflito real encontrado na auditoria e corrigido, não deixado de lado:
Organization.treasury (Fase 13J) era Float, mutado por acumulação direta
de ponto flutuante (organization.treasury += delta) — exatamente o
"floating-point money" que a 14A proíbe explicitamente. A própria
docstring da migração da 13J já tinha deixado isso pra "Phase 14
resolver". Migração converte a coluna de Float pra Integer (Bronze, a
unidade canônica) via batch_alter_table; deposit_funds/withdraw_funds/
_change_funds (13J) agora recebem int; os testes da 13J que usavam float
foram atualizados.

- CurrencyHolding (novo): dinheiro fisicamente detido, em Bronze — nunca
  uma linha por moeda. owner_type reaproveita CombatActorType (mesmo
  vocabulário já usado por Group/OrganizationMember na Fase 13).
  container_item_instance_id opcional reaproveita um ItemInstance real da
  Fase 10 (um baú) pra dinheiro guardado num lugar físico, sem precisar
  de um item "carteira"; None significa carregado pessoalmente. Dinheiro
  de organização continua em Organization.treasury (agora inteiro) — não
  foi unificado no CurrencyHolding, que é só pra CHARACTER/NPC/
  SIMULATED_PLAYER.
- app/game/economy/currency.py: matemática pura de conversão
  (to_denominations/from_denominations), 100 Bronze = 1 Silver, 100
  Silver = 1 Gold fixos, inteiro em todo lugar. Valores negativos e não
  inteiros levantam CurrencyError explicitamente em vez de aceitos por
  acidente.
- app/game/economy/wallet.py: deposit/withdraw/transfer (primitiva crua
  de movimentação, sem preço/item — isso é a 14C) e
  total_carried_by_owner (soma de conveniência entre todas as holdings).

Decisão deliberada: ItemType.CURRENCY já existia (vocabulário morto,
nunca usado) mas NÃO foi reaproveitado para empilhar moedas como
ItemInstance — arrastaria qualidade, durabilidade, slot de equipamento e
peso, tudo irrelevante pra dinheiro, e fragmentaria o saldo em várias
linhas por combinação de material/qualidade.

21 testes novos (incluindo os limites exigidos pela spec: zero, 99/100
Bronze, 99/100 Silver, valores grandes, negativo explicitamente rejeitado)
+ 4 testes da 13J atualizados. Suíte completa: 925/925 backend.

+++ backend/app/db/models/organization.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Float, ForeignKey, Integer, String, UniqueConstraint
+from sqlalchemy import ForeignKey, Integer, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.enums import (
@@ -48,7 +48,9 @@ class Organization(Base):
         String, default=OrganizationFormality.INFORMAL, nullable=False
     )
     founding_group_id: Mapped[str | None] = mapped_column(ForeignKey("groups.id"), nullable=True)
-    treasury: Mapped[float] = mapped_column(Float, default=0.0, nullable=False)
+    # In Bronze (Phase 14A's canonical smallest unit) — was a Float (Phase
+    # 13J) until Phase 14A established money is always integer.
+    treasury: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
 
 
 class OrganizationRole(Base):
